fix(circle ci configs): update circle-ci configs (#299)
